backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
langgraph
faiss-cpu
sentence-transformers
numpy